BE/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiomysql==0.1.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.3.0
bcrypt==3.2.2
bitarray==3.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ckzg==2.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.0
cytoolz==1.0.1
dateparser==1.2.1
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
dnspython==2.6.1
dynaconf==3.2.6
email_validator==2.2.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.112.1
fastapi-cli==0.0.5
fastapi-sa==0.1.0
frozenlist==1.5.0
greenlet==3.0.3
h11==0.14.0
hexbytes==1.3.0
hiredis==3.0.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.7
inflection==0.5.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.65.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.1
pyunormalize==16.0.0
PyYAML==6.0.2
pyzmq==26.3.0
redis==5.0.8
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rlp==4.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
stack-data==0.6.3
starlette==0.38.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.12.4
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.2.2
uuid_utils==0.9.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==0.23.0
wcwidth==0.2.13
web3==7.9.0
websockets==13.0
yarl==1.18.3
